Feat: plan 날짜별로 불러오기

+++ back-end/backend/src/main/java/com/groot/backend/repository/PlanRepositoryCustom.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface PlanRepositoryCustom {
     List<PlanEntity> findAllByDateTime(LocalDateTime start, LocalDateTime end);
 
-    int updateDoneAndDateTimeByCodeAndPotId(Integer code, Long potId);
+    long updateDoneAndDateTimeByCodeAndPotId(Integer code, Long potId);
 }
